chg: cleanup: Ensure not following sysmlinks
when removing or compressing files.

pen-torproject
Closes #5

+++ sbws/core/cleanup.py
@@ -54,6 +54,8 @@ def _get_files_mtime_older_than(dname, days_delta, extensions):
     # Determine oldest allowed date
     today = datetime.utcfromtimestamp(time.time())
     oldest_day = today - timedelta(days=days_delta)
+    # By default, `os.walk`` doesn't follow symlinks.
+    # (https://docs.python.org/3/library/os.html#os.walk)
     for root, dirs, files in os.walk(dname):
         for f in files:
             fname = os.path.join(root, f)
@@ -67,6 +69,8 @@ def _get_files_mtime_older_than(dname, days_delta, extensions):
                 continue
             # using file modification time instead of parsing the name
             # of the file.
+            # `os.stat` follows sysmlinks by default
+            # (https://docs.python.org/3/library/os.html#os.stat)
             filedt = unixts_to_dt_obj(
                 os.stat(fname, follow_symlinks=False).st_mtime
             )
@@ -82,7 +86,9 @@ def _delete_files(dname, files, dry_run=True):
         for fname in files:
             log.info("Deleting %s", fname)
             assert os.path.commonprefix([dname, fname]) == dname
-            if not dry_run:
+            # Ensure fname isn't a symlink even if `files` are obtained via
+            # `os.walk`.
+            if not dry_run and not os.path.islink(fname):
                 os.remove(fname)
 
 
@@ -94,7 +100,9 @@ def _compress_files(dname, files, dry_run=True):
         for fname in files:
             log.info("Compressing %s", fname)
             assert os.path.commonprefix([dname, fname]) == dname
-            if dry_run:
+            # Ensure fname isn't a symlink even if `files` are obtained via
+            # `os.walk`.
+            if dry_run or os.path.islink(fname):
                 continue
             with open(fname, "rt") as in_fd:
                 out_fname = fname + ".gz"
